Refactor recreate collection (#719)
* deprecated recreate_collection from tests

* checked if collection exists before deletion

* reverted deletion of recreate_collection

* added missing collection_exsits before deletion

* nit if condition

* fix type hint in local_collection

* removed run mypy on async client generator

* added checks for create_collection if exists

* nit change collection name

* add else to create_collection

* nit

* nit

* refactoring: remove redundant checks

* fix: do not test collection_exists through itself

* refactoring: remove redundant collection exists checks

* fix: remove redundant collection exists checks

* refactoring: remove redundant collection checks

* refactoring: remove redundant collection existence check in test sparse search

* refactoring: remove redundant checks in test_updates

* refactoring: remove redundant checks

* fix: remove redundant checks

* refactoring: remove redundant checks

* refactoring: remove redundant imports

* new: add recreate collection test

---------

+++ tests/test_async_qdrant_client.py
@@ -113,8 +113,6 @@ async def test_async_qdrant_client(prefer_grpc):
         await client.delete_collection(COLLECTION_NAME)
         await client.create_collection(**collection_params)
 
-    await client.recreate_collection(**collection_params)
-
     await client.get_collection(COLLECTION_NAME)
     await client.get_collections()
     if not version_set or dev_version or minor_version >= 8:
@@ -355,9 +353,9 @@ async def test_async_qdrant_client_local():
         collection_name=COLLECTION_NAME,
         vectors_config=models.VectorParams(size=10, distance=models.Distance.EUCLID),
     )
+    if await client.collection_exists(COLLECTION_NAME):
+        await client.delete_collection(COLLECTION_NAME)
     await client.create_collection(**collection_params)
-    await client.delete_collection(COLLECTION_NAME)
-    await client.recreate_collection(**collection_params)
 
     await client.get_collection(COLLECTION_NAME)
     await client.get_collections()
@@ -537,7 +535,8 @@ async def test_async_qdrant_client_local():
     assert (await client.retrieve(COLLECTION_NAME, ids=[1]))[0].payload == {}
 
     # region teardown
-    await client.delete_collection(COLLECTION_NAME)
+    if await client.collection_exists(COLLECTION_NAME):
+        await client.delete_collection(COLLECTION_NAME)
     collections = await client.get_collections()
 
     assert all(collection.name != COLLECTION_NAME for collection in collections.collections)
