+++ tests/test_qdrant_client.py
@@ -6,8 +6,10 @@ from tempfile import mkdtemp
 from time import sleep
 from typing import List
 
+
 import numpy as np
 import pytest
+from httpx import Timeout
 from grpc import Compression, RpcError
 
 import qdrant_client.http.exceptions
@@ -199,8 +201,9 @@ def test_records_upload(prefer_grpc, parallel):
     )
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -234,7 +237,8 @@ def test_records_upload(prefer_grpc, parallel):
 
     records = (Record(id=idx, vector=np.random.rand(DIM).tolist()) for idx in range(NUM_VECTORS))
 
-    client.recreate_collection(
+    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -263,7 +267,9 @@ def test_point_upload(prefer_grpc, parallel):
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
 
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -295,7 +301,8 @@ def test_point_upload(prefer_grpc, parallel):
     assert result_count.count < 900
     assert result_count.count > 100
 
-    client.recreate_collection(
+    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -321,7 +328,9 @@ def test_upload_collection(prefer_grpc, parallel):
     batch_size = 2
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
 
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=size, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -346,7 +355,8 @@ def test_upload_collection(prefer_grpc, parallel):
 
     assert client.get_collection(collection_name=COLLECTION_NAME).points_count == 5
 
-    client.recreate_collection(
+    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=size, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -381,8 +391,9 @@ def test_multiple_vectors(prefer_grpc):
     ]
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={
             "image": VectorParams(size=DIM, distance=Distance.DOT),
@@ -440,8 +451,9 @@ def test_qdrant_client_integration(prefer_grpc, numpy_upload, local_mode):
         client = QdrantClient(location=":memory:", prefer_grpc=prefer_grpc)
     else:
         client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -1032,8 +1044,9 @@ def test_qdrant_client_integration(prefer_grpc, numpy_upload, local_mode):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_qdrant_client_integration_update_collection(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={
             "text": VectorParams(size=DIM, distance=Distance.DOT),
@@ -1092,8 +1105,9 @@ def test_qdrant_client_integration_update_collection(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_points_crud(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -1141,8 +1155,9 @@ def test_points_crud(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_quantization_config(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         quantization_config=ScalarQuantization(
@@ -1201,7 +1216,9 @@ def test_custom_sharding(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
 
     def init_collection():
-        client.recreate_collection(
+        if client.collection_exists(COLLECTION_NAME):
+            client.delete_collection(collection_name=COLLECTION_NAME)
+        client.create_collection(
             collection_name=COLLECTION_NAME,
             vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
             sharding_method=models.ShardingMethod.CUSTOM,
@@ -1375,8 +1392,9 @@ def test_sparse_vectors(prefer_grpc):
         pytest.skip("Sparse vectors are supported since v1.7.0")
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={},
         sparse_vectors_config={
@@ -1454,8 +1472,9 @@ def test_sparse_vectors_batch(prefer_grpc):
         pytest.skip("Sparse vectors are supported since v1.7.0")
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={},
         sparse_vectors_config={
@@ -1536,8 +1555,9 @@ def test_sparse_vectors_batch(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_vector_update(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -1595,8 +1615,9 @@ def test_vector_update(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_conditional_payload_update(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -1634,8 +1655,9 @@ def test_conditional_payload_update(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_conditional_payload_update_2(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -1714,8 +1736,9 @@ def test_insert_float():
 
 def test_locks():
     client = QdrantClient(timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
         timeout=TIMEOUT,
@@ -1760,8 +1783,9 @@ def test_locks():
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_empty_vector(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.recreate_collection(
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={},
         timeout=TIMEOUT,
@@ -1820,19 +1844,25 @@ def test_client_close():
 
     # region http
     client_http = QdrantClient(timeout=TIMEOUT)
-    client_http.recreate_collection(
+    if client_http.collection_exists("test"):
+        client_http.delete_collection("test")
+    client_http.create_collection(
         "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
     )
     client_http.close()
     with pytest.raises(qdrant_exceptions.ResponseHandlingException):
-        client_http.recreate_collection(
+        if client_http.collection_exists("test"):
+            client_http.delete_collection("test")
+        client_http.create_collection(
             "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
         )
     # endregion
 
     # region grpc
     client_grpc = QdrantClient(prefer_grpc=True, timeout=TIMEOUT)
-    client_grpc.recreate_collection(
+    if client_grpc.collection_exists("test"):
+        client_grpc.delete_collection("test")
+    client_grpc.create_collection(
         "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
     )
     client_grpc.close()
@@ -1873,7 +1903,9 @@ def test_client_close():
 
     # region local
     local_client_in_mem = QdrantClient(":memory:")
-    local_client_in_mem.recreate_collection(
+    if local_client_in_mem.collection_exists("test"):
+        local_client_in_mem.delete_collection("test")
+    local_client_in_mem.create_collection(
         "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
     )
     local_client_in_mem.close()
@@ -1885,24 +1917,30 @@ def test_client_close():
         )
 
     with pytest.raises(RuntimeError):
-        local_client_in_mem.create_collection(
-            "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
-        )
+        if not local_client_in_mem.collection_exists("test"):
+            local_client_in_mem.create_collection(
+                "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
+            )
 
     with pytest.raises(RuntimeError):
-        local_client_in_mem.delete_collection("test")
+        if local_client_in_mem.collection_exists("test"):
+            local_client_in_mem.delete_collection("test")
 
     with tempfile.TemporaryDirectory() as tmpdir:
         path = tmpdir + "/test.db"
 
         local_client_persist_1 = QdrantClient(path=path)
-        local_client_persist_1.recreate_collection(
+        if local_client_persist_1.collection_exists("test"):
+            local_client_persist_1.delete_collection("test")
+        local_client_persist_1.create_collection(
             "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
         )
         local_client_persist_1.close()
 
         local_client_persist_2 = QdrantClient(path=path)
-        local_client_persist_2.recreate_collection(
+        if local_client_persist_2.collection_exists("test"):
+            local_client_persist_2.delete_collection("test")
+        local_client_persist_2.create_collection(
             "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
         )
         local_client_persist_2.close()
@@ -1910,10 +1948,6 @@ def test_client_close():
 
 
 def test_timeout_propagation():
-    import time
-
-    from httpx import Timeout
-
     client = QdrantClient()
     vectors_config = models.VectorParams(size=2, distance=models.Distance.COSINE)
     with pytest.raises(
@@ -1922,9 +1956,13 @@ def test_timeout_propagation():
         # timeout is Optional[int]
         # if we set it to 0 - recreate_collection raises operation is in progress instead of timed out
         client.http.client._client._timeout = Timeout(0.01)
-        client.recreate_collection(collection_name=COLLECTION_NAME, vectors_config=vectors_config)
-    time.sleep(0.5)
-    client.recreate_collection(
+        if client.collection_exists(COLLECTION_NAME):
+            client.delete_collection(collection_name=COLLECTION_NAME)
+        client.create_collection(collection_name=COLLECTION_NAME, vectors_config=vectors_config)
+    sleep(0.5)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=10)
+    client.create_collection(
         collection_name=COLLECTION_NAME, vectors_config=vectors_config, timeout=10
     )
 
@@ -1937,10 +1975,11 @@ def test_grpc_options():
     assert client._client._grpc_options == {"grpc.max_send_message_length": 3}
 
     with pytest.raises(RpcError):
-        client.create_collection(
-            "grpc_collection",
-            vectors_config=models.VectorParams(size=100, distance=models.Distance.COSINE),
-        )
+        if not client.collection_exists("grpc_collection"):
+            client.create_collection(
+                "grpc_collection",
+                vectors_config=models.VectorParams(size=100, distance=models.Distance.COSINE),
+            )
 
 
 def test_grpc_compression():
